backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lightgbm==4.6.0
loky==3.4.1
MarkupSafe==3.0.2
numpy==2.2.3
pandas==2.2.3
pygame==2.6.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
Werkzeug==3.1.3
xgboost==2.1.4
